Sleep delay option, cancellable countdown, UI non-locking timer
Added command line option to specify the delay before it puts your monitors to sleep.
"-sleepDelaySeconds 5" as an example.

Can now press ESC to cancel the program if you change your mind or clicked by mistake :)

Implemented a countdown timer that will also show on the tiny pop-up when in non-silent mode.